+++ testing/validate_task_parms_Monthly.cpp
@@ -54,7 +54,7 @@ int test1(){
         hours + ":" + minutes + ":" + seconds, // HH:MM:SS
         years + "-" + months + "-" + days + " " + hours + ":" + minutes + ":" + seconds, // YYYY-MM-DD HH:MM:SS
     };
-    for(int i = 0; i < datetimes.size(); i++){
+    for(size_t i = 0; i < datetimes.size(); i++){
         cl::Config* c = new cl::Config();
         c->add_entry("Name", "Test Title");
         c->add_entry("Description", "A short description");
@@ -215,7 +215,7 @@ int test3(){
         wday_abbr + " " + hours + ":" + minutes + ":" + seconds, // WDAY HH:MM:SS (abbreviated week day name)
         wday_full + " " + hours + ":" + minutes + ":" + seconds // WDAY HH:MM:SS (full week day name)
     };
-    for(int i = 0; i < datetimes.size(); i++){
+    for(size_t i = 0; i < datetimes.size(); i++){
         cl::Config* c = new cl::Config();
         c->add_entry("Name", "Test Title");
         c->add_entry("Description", "A short description");
@@ -251,7 +251,7 @@ int test4(){
         {"12:11111:00", ts::ValidationCode::BAD_HMS_LENGTH},
         {"12:00:11111", ts::ValidationCode::BAD_HMS_LENGTH},
     };
-    for(int i = 0; i < datetimes_retcodes.size(); i++){
+    for(size_t i = 0; i < datetimes_retcodes.size(); i++){
         cl::Config* c = new cl::Config();
         c->add_entry("Name", "Test Title");
         c->add_entry("Description", "A short description");
@@ -294,7 +294,7 @@ int test5(){
         {"2022-2-15 12:00:00", ts::ValidationCode::BAD_YYYYMMDD_LENGTH},
     };
 
-    for(int i = 0; i < datetimes_retcodes.size(); i++){
+    for(size_t i = 0; i < datetimes_retcodes.size(); i++){
         cl::Config* c = new cl::Config();
         c->add_entry("Name", "Test Title");
         c->add_entry("Description", "A short description");
@@ -334,7 +334,7 @@ int test6(){
         {"2022-2-15", ts::ValidationCode::BAD_YYYYMMDD_LENGTH},
     };
 
-    for(int i = 0; i < datetimes_retcodes.size(); i++){
+    for(size_t i = 0; i < datetimes_retcodes.size(); i++){
         cl::Config* c = new cl::Config();
         c->add_entry("Name", "Test Title");
         c->add_entry("Description", "A short description");
@@ -427,7 +427,7 @@ int test8(){
         hours + ":" + minutes + ":" + seconds, // HH:MM:SS
         years + "-" + months + "-" + days + " " + hours + ":" + minutes + ":" + seconds, // YYYY-MM-DD HH:MM:SS
     };
-    for(int i = 0; i < datetimes.size(); i++){
+    for(size_t i = 0; i < datetimes.size(); i++){
         cl::Config* c = new cl::Config();
         c->add_entry("Name", "Test Title");
         c->add_entry("Description", "A short description");
